Consensus engine refactor
