+++ src/chain/mod.rs
@@ -1,8 +1,5 @@
-pub mod blockchain;
 pub mod config;
-pub mod consensus;
 pub mod dao;
 pub mod difficulty;
 pub mod intrinsic_gas;
 pub mod protocol_param;
-pub mod validity;
