+++ src/lib.rs
@@ -22,6 +22,7 @@ pub mod adapter;
 mod bitmapdb;
 pub mod chain;
 mod changeset;
+pub mod consensus;
 pub mod crypto;
 pub mod downloader;
 pub mod etl;
